requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
sentence_transformers
langchain_huggingface
faiss-cpu
chromadb
langchain_chroma
streamlit
arxiv
wikipedia
duckduckgo-search
google-search-results
sqlalchemy
mysql-connector-python
validators==0.28.1
youtube_transcript_api
pytube
unstructured
langchain-google-genai
numexpr